Move code into new signup activity.
